Add books_manager file
